Ace CodeDex 看板告警清零，涉及计napi、ace_engin相关库

+++ frameworks/base/i18n/localization.cpp
@@ -566,11 +566,12 @@ LunarDate Localization::GetLunarDate(Date date)
     CHECK_RETURN(status, dateRet);
 
     // Sexagenary cycle years convert to Western years
-    dateRet.year = lunarYear + GUIHAI_YEAR_RECENT;
-    dateRet.year += ((date.year - GUIHAI_YEAR_RECENT) / SEXAGENARY_CYCLE_SIZE) * SEXAGENARY_CYCLE_SIZE;
+    dateRet.year = static_cast<uint32_t>(lunarYear) + GUIHAI_YEAR_RECENT;
+    dateRet.year +=
+        ((static_cast<uint32_t>(date.year) - GUIHAI_YEAR_RECENT) / SEXAGENARY_CYCLE_SIZE) * SEXAGENARY_CYCLE_SIZE;
     // 0 means January,  1 means February, so month + 1
-    dateRet.month = lunarMonth + 1;
-    dateRet.day = lunarDate;
+    dateRet.month = static_cast<uint32_t>(lunarMonth) + 1;
+    dateRet.day = static_cast<uint32_t>(lunarDate);
     dateRet.isLeapMonth = isLeapMonth == 0 ? false : true;
     return dateRet;
 }
